move flo client library into separate crate
