+++ Cargo.toml
@@ -6,6 +6,7 @@ authors = ["pfried <philipsfried@gmail.com>"]
 [dependencies]
 flo-event = { path = "./flo-event" }
 flo-protocol = { path = "./flo-protocol" }
+flo-client-lib = { path = "./flo-client-lib" }
 num_cpus = "^1.2"
 log = "0.3"
 log4rs = "0.4"
@@ -29,7 +30,3 @@ path = "src/main.rs"
 [[bin]]
 name = "flo-client"
 path = "src/flo_client_cli.rs"
-
-[lib]
-name = "flo_sync_client"
-path = "src/sync_client.rs"
